+++ src/app/[lang]/(protected)/project/[id]/page.tsx
@@ -3,16 +3,14 @@ import ProjectTabs from "@/components/Project/Tabs";
 import MockData from "./project.json";
 
 export default async function Project() {
-  const project = MockData;
-
   return (
     <TimelineProvider>
       <main className="container flex flex-col gap-6 py-10 items-center lg:items-start w-full">
         <ProjectTabs
-          project={project}
+          project={MockData}
           contractTime={null}
           contractId="contractTestId"
-          contractHistories={project.contractHistories}
+          contractHistories={MockData.contractHistories}
         />
       </main>
     </TimelineProvider>
