+++ src/components/Project/ContractHistoryTable.tsx
@@ -17,30 +17,28 @@ interface ContractHistoryTableProps {
 
 const ContractHistoryTable: React.FC<ContractHistoryTableProps> = ({
   contractHistories,
-}) => {
-  return (
-    <Table>
-      <TableCaption>
-        {contractHistories.length ? "List of Signers" : "No signer yet"}
-      </TableCaption>
-      <TableHeader>
-        <TableRow>
-          <TableHead>Collaborator</TableHead>
-          <TableHead>Status</TableHead>
+}) => (
+  <Table>
+    <TableCaption>
+      {contractHistories.length ? "List of Signers" : "No signer yet"}
+    </TableCaption>
+    <TableHeader>
+      <TableRow>
+        <TableHead>Collaborator</TableHead>
+        <TableHead>Status</TableHead>
+      </TableRow>
+    </TableHeader>
+    <TableBody>
+      {contractHistories.map((history, id: number) => (
+        <TableRow key={id.toString()}>
+          <TableCell>{history.projectUser?.user_name}</TableCell>
+          <TableCell>
+            Signed at {new Date(history.created_at).toLocaleString()}
+          </TableCell>
         </TableRow>
-      </TableHeader>
-      <TableBody>
-        {contractHistories.map((history, id: number) => (
-          <TableRow key={id.toString()}>
-            <TableCell>{history.projectUser?.user_name}</TableCell>
-            <TableCell>
-              Signed at {new Date(history.created_at).toLocaleString()}
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>
-  );
-};
+      ))}
+    </TableBody>
+  </Table>
+);
 
 export default ContractHistoryTable;
